update to clap 3.1.0, replace deprecated code
