+++ Cargo.toml
@@ -20,7 +20,7 @@ libc = "0.2"
 bytes = "1"
 socket2 = "0.4"
 
-clap = "3"
+clap = "3.1"
 toml = "0.5"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
